Youtube API V2 First Commit
Version 2 revision from version 1 includes loading data via PHP and setting the data to local storage to better secure the API key.  Cookie is also generated with a timestamp to avoid calling from the API too many times.  Javascript is used afterward to render video items and lightbox on the page utilizing the data that was saved in local storage.